Extract tool config into codex-tools
