+++ codex-rs/tools/Cargo.toml
@@ -10,7 +10,10 @@ workspace = true
 [dependencies]
 codex-app-server-protocol = { workspace = true }
 codex-code-mode = { workspace = true }
+codex-features = { workspace = true }
 codex-protocol = { workspace = true }
+codex-utils-absolute-path = { workspace = true }
+codex-utils-pty = { workspace = true }
 rmcp = { workspace = true, default-features = false, features = [
     "base64",
     "macros",
@@ -19,6 +22,7 @@ rmcp = { workspace = true, default-features = false, features = [
 ] }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
+tracing = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
