+++ codex-rs/tools/src/lib.rs
@@ -5,6 +5,7 @@ mod agent_job_tool;
 mod agent_tool;
 mod code_mode;
 mod dynamic_tool;
+mod image_detail;
 mod js_repl_tool;
 mod json_schema;
 mod local_tool;
@@ -12,6 +13,7 @@ mod mcp_resource_tool;
 mod mcp_tool;
 mod request_user_input_tool;
 mod responses_api;
+mod tool_config;
 mod tool_definition;
 mod tool_discovery;
 mod tool_spec;
@@ -38,6 +40,8 @@ pub use code_mode::create_code_mode_tool;
 pub use code_mode::create_wait_tool;
 pub use code_mode::tool_spec_to_code_mode_tool_definition;
 pub use dynamic_tool::parse_dynamic_tool;
+pub use image_detail::can_request_original_image_detail;
+pub use image_detail::normalize_output_image_detail;
 pub use js_repl_tool::create_js_repl_reset_tool;
 pub use js_repl_tool::create_js_repl_tool;
 pub use json_schema::AdditionalProperties;
@@ -66,6 +70,12 @@ pub use responses_api::dynamic_tool_to_responses_api_tool;
 pub use responses_api::mcp_tool_to_deferred_responses_api_tool;
 pub use responses_api::mcp_tool_to_responses_api_tool;
 pub use responses_api::tool_definition_to_responses_api_tool;
+pub use tool_config::ShellCommandBackendConfig;
+pub use tool_config::ToolUserShellType;
+pub use tool_config::ToolsConfig;
+pub use tool_config::ToolsConfigParams;
+pub use tool_config::UnifiedExecShellMode;
+pub use tool_config::ZshForkConfig;
 pub use tool_definition::ToolDefinition;
 pub use tool_discovery::DiscoverablePluginInfo;
 pub use tool_discovery::DiscoverableTool;
